Rename TheActivityLifecycle -> ActivityLifecycle
