+++ app/src/main/java/org/simple/clinic/activity/ActivityLifecycle.kt
@@ -4,25 +4,25 @@ import org.simple.clinic.widgets.UiEvent
 
 private fun event(activityName: String?, event: String): String = activityName?.let { "$activityName:$event" } ?: ""
 
-sealed class TheActivityLifecycle : UiEvent {
+sealed class ActivityLifecycle : UiEvent {
 
-  class Resumed(activityName: String?) : TheActivityLifecycle() {
+  class Resumed(activityName: String?) : ActivityLifecycle() {
     override val analyticsName = event(activityName, "Resumed")
   }
 
-  class Started(activityName: String?) : TheActivityLifecycle() {
+  class Started(activityName: String?) : ActivityLifecycle() {
     override val analyticsName = event(activityName, "Started")
   }
 
-  class Paused(activityName: String?) : TheActivityLifecycle() {
+  class Paused(activityName: String?) : ActivityLifecycle() {
     override val analyticsName = event(activityName, "Paused")
   }
 
-  class Stopped(activityName: String?) : TheActivityLifecycle() {
+  class Stopped(activityName: String?) : ActivityLifecycle() {
     override val analyticsName = event(activityName, "Stopped")
   }
 
-  class Destroyed(activityName: String?) : TheActivityLifecycle() {
+  class Destroyed(activityName: String?) : ActivityLifecycle() {
     override val analyticsName = event(activityName, "Destroyed")
   }
 }
